refactor(scripting): move the script module files into their own directory
Two files carrying the same long prefix read better as a directory. The
exported names keep theirs — ScriptModuleArtifact and storeScriptModule say
what they are at a call site, where the directory is not visible.
